requirements.txt
appdirs==1.4.3
Django==1.11.2
django-suit==0.2.25
django-suit-redactor==0.0.4
django-widget-tweaks==1.4.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
virtualenv==15.1.0
